+++ server/index.ts
@@ -5,9 +5,8 @@ import { Socket } from 'socket.io'
 import { v4 as uuid } from 'uuid'
 
 import { ERROR_NOTICE } from '../constants/chat.js'
-import { gameConfig, PRICE_THRESHOLD } from '../constants/game.js'
+import { gameConfig } from '../constants/game.js'
 import { loadGameHints } from '../lib/api/hints.js'
-import { generateNewFishPrice, isPriceChangeHigh, shouldHintMatch } from '../lib/utils/game.js'
 import { generateGameId } from '../lib/utils/generate-game-id.js'
 import { GameModel, PlayerModel } from '../types/game.js'
 
@@ -15,7 +14,6 @@ import {
   gameRooms,
   playersMap,
   gameTimers,
-  roundTimers,
   getRoom,
   addRoom,
   removeRoom,
@@ -23,6 +21,7 @@ import {
   removePlayer,
   getPlayerId,
 } from './game/store.js'
+import { clearGameTimers, startGameTimer } from './game/timer.js'
 import { createSocketServer } from './socket/config.js'
 
 const dev = process.env.NODE_ENV !== 'production'
@@ -35,158 +34,6 @@ const handler = app.getRequestHandler()
 app.prepare().then(() => {
   const httpServer = createServer(handler)
   const io = createSocketServer(httpServer)
-
-  const startGameTimer = (gameId: string, room: GameModel & { readyPlayers: Set<string> }) => {
-    if (!room || room.state !== 'in_progress') {
-      console.error('게임을 시작할 수 없는 상태입니다')
-      return
-    }
-
-    if (room.gameInfo.currentDay < 1 || room.gameInfo.currentDay > room.totalRounds) {
-      console.error('유효하지 않은 라운드입니다')
-      room.gameInfo.currentDay = Math.max(1, Math.min(room.totalRounds, room.gameInfo.currentDay))
-    }
-
-    const existingTimer = roundTimers.get(gameId)
-    if (existingTimer) {
-      clearInterval(existingTimer)
-      roundTimers.delete(gameId)
-    }
-
-    let timer = gameConfig.INITIAL_TIMER
-    const intervalId = setInterval(() => {
-      const currentRoom = getRoom(gameId)
-      if (!currentRoom || currentRoom.state !== 'in_progress') {
-        clearInterval(intervalId)
-        roundTimers.delete(gameId)
-        return
-      }
-
-      timer -= 1
-      io.to(gameId).emit('timer_update', timer)
-
-      if (timer === 0) {
-        timer = gameConfig.INITIAL_TIMER
-        io.to(gameId).emit('timer_update', timer)
-        clearInterval(intervalId)
-        roundTimers.delete(gameId)
-
-        try {
-          const isLastRound = currentRoom.gameInfo.currentDay === currentRoom.totalRounds
-
-          if (isLastRound) {
-            try {
-              const prevHint = currentRoom.hints[currentRoom.gameInfo.currentDay - 2]
-              if (!prevHint) return
-
-              const isHintMatched = shouldHintMatch()
-              const priceChangeDirection = isHintMatched
-                ? prevHint.expectedChange
-                : prevHint.expectedChange === 'up'
-                  ? 'down'
-                  : 'up'
-
-              const oldPrice = currentRoom.gameInfo.currentFishPrice
-              const newPrice = generateNewFishPrice(oldPrice, priceChangeDirection)
-
-              io.to(gameId).emit('last_fish_price', newPrice)
-            } catch (error) {
-              console.error('마지막 라운드의 생선 가격을 계산하는데 실패하였습니다.', error)
-            }
-          } else {
-            try {
-              if (currentRoom.gameInfo.currentDay >= currentRoom.totalRounds) {
-                console.error('현재 라운드가 전체 라운드 수를 초과하였습니다.')
-                return
-              }
-
-              currentRoom.gameInfo.currentDay += 1
-
-              if (currentRoom.gameInfo.currentDay > currentRoom.totalRounds) {
-                console.error('라운드 수가 전체 라운드 수를 초과하였습니다.')
-                currentRoom.gameInfo.currentDay = currentRoom.totalRounds
-              }
-
-              const currentPrice = currentRoom.gameInfo.currentFishPrice
-
-              let isHintMatched = shouldHintMatch()
-              const currentHint = currentRoom.hints[currentRoom.gameInfo.currentDay - 1]
-              const prevHint = currentRoom.hints[currentRoom.gameInfo.currentDay - 2]
-
-              if (
-                currentPrice <= PRICE_THRESHOLD.MIN_SAFE_PRICE &&
-                prevHint?.expectedChange === 'down'
-              ) {
-                isHintMatched = false
-              } else if (
-                currentPrice >= PRICE_THRESHOLD.MAX_SAFE_PRICE &&
-                prevHint?.expectedChange === 'up'
-              ) {
-                isHintMatched = false
-              }
-
-              const priceChangeDirection = isHintMatched
-                ? prevHint.expectedChange
-                : prevHint.expectedChange === 'up'
-                  ? 'down'
-                  : 'up'
-
-              const oldPrice = currentRoom.gameInfo.currentFishPrice
-              const newPrice = generateNewFishPrice(oldPrice, priceChangeDirection)
-
-              const isHighChange = isPriceChangeHigh(oldPrice, newPrice)
-              let outcomeMessage = ''
-
-              if (isHintMatched) {
-                outcomeMessage = isHighChange ? prevHint.matchOutcomeHigh : prevHint.matchOutcomeLow
-              } else {
-                outcomeMessage = isHighChange
-                  ? prevHint.mismatchOutcomeHigh
-                  : prevHint.mismatchOutcomeLow
-              }
-
-              currentRoom.gameInfo = {
-                ...currentRoom.gameInfo,
-                currentFishPrice: newPrice,
-                lastRoundHintResult: outcomeMessage,
-                nextRoundHint: currentHint?.hint || '',
-              }
-
-              io.to(gameId).emit('update_game_info', currentRoom.gameInfo)
-
-              startGameTimer(gameId, currentRoom)
-            } catch (error) {
-              console.error('다음 라운드로 게임을 업데이트하는데 실패하였습니다.', error)
-              if (currentRoom) {
-                io.to(gameId).emit('update_game_info', currentRoom.gameInfo)
-              }
-            }
-          }
-        } catch (error) {
-          console.error('Round update error:', error)
-          if (currentRoom) {
-            io.to(gameId).emit('update_game_info', currentRoom.gameInfo)
-          }
-        }
-      }
-    }, 1000)
-
-    roundTimers.set(gameId, intervalId)
-  }
-
-  const clearGameTimers = (gameId: string) => {
-    const gameTimer = gameTimers.get(gameId)
-    if (gameTimer) {
-      clearTimeout(gameTimer)
-      gameTimers.delete(gameId)
-    }
-
-    const roundTimer = roundTimers.get(gameId)
-    if (roundTimer) {
-      clearInterval(roundTimer)
-      roundTimers.delete(gameId)
-    }
-  }
 
   const removePlayerFromRoom = (
     socket: Socket,
@@ -401,7 +248,7 @@ app.prepare().then(() => {
       room.readyPlayers.add(playerId)
 
       if (room.readyPlayers.size === room.players.length) {
-        startGameTimer(gameId, room)
+        startGameTimer(io, gameId, room)
         io.to(gameId).emit('all_players_ready')
       }
     })
